Update NPM packages
Includes a fix for CVE-2022-31160 in jQuery UI. This vulnerability does
not seem to impact webchat because we don't use radio buttons.

Includes fixes for CVE-2022-37599, CVE-2022-37601 and CVE-2022-37603 in
webpack loader-utils. These vulnerabilities does not seem to impact
webchat because we don't run webpack on untrusted input.

Includes a fix for CVE-2022-3517 in minimatch. This vulnerability does
not seem to impact webchat because we don't run it on untrusted input.

Includes a fix for CVE-2021-44906 in minimist. This vulnerability does
not seem to impact webchat because we don't run it on untrusted input.

Includes a fix for CVE-2022-25758 in scss-tokenizer. This vulnerability
does not seem to impact webchat because we don't run it on untrusted
input.

Inlcude s a fix for CVE-2022-25858 in terser. This vulnerability does
not seem to impact webchat because we don't run it on untrusted input.